requirements.txt
absl-py==0.8.1
aiohttp==3.5.4
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
APScheduler==3.6.1
asn1crypto==1.2.0
astor==0.8.0
astroid==2.2.5
astropy==3.2.2
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.4
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.1
bitarray==1.0.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.13.0
chardet==3.0.4
circlify==0.11.0
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.8.3
conda-build==3.18.10
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.6.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
discord==1.0.1
discord.py==1.2.3
distributed==2.6.0
Django==2.2.4
django-admin-list-filter-dropdown==1.0.2
django-admin-rangefilter==0.5.0
django-anymail==7.0.0
django-multiselectfield==0.1.8
django-phonenumber-field==3.0.1
django-storages==1.7.2
djangorestframework==3.10.2
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.7.8
Flask==1.1.1
Flask-Cors==3.0.8
fsspec==0.5.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.7
google-api-core==1.14.3
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-pasta==0.1.7
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.3
gunicorn==19.9.0
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
htmlmin==0.1.12
idna==2.8
idna-ssl==1.1.0
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
jsmin==2.2.2
json5==0.8.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
livereload==2.6.1
llvmlite==0.30.0
locket==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkdocs==1.0.4
mkdocs-material==4.6.0
mkdocs-minify-plugin==0.2.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.5.2
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.46.0
numexpr==2.7.0
numpy==1.17.3
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
opencv-python==4.1.1.26
openpyxl==3.0.0
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pbr==5.4.3
pep562==1.0
pep8==1.7.1
phonenumbers==8.10.15
pickleshare==0.7.5
Pillow==6.2.0
pipenv==2020.8.13
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
praw==6.4.0
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.3
psycopg2-binary==2.8.3
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.4.2
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pymdown-extensions==6.2
pymongo==3.10.1
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pypng==0.0.20
PyQRCode==1.2.1
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-dotenv==0.13.0
pytz==2019.2
PyWavelets==1.1.1
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spotipy==2.12.0
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.10
sqlparse==0.3.0
statsmodels==0.10.1
stevedore==1.31.0
sympy==1.4
tables==3.5.2
tblib==1.5.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
typing==3.6.4
typing-extensions==3.7.4
tzlocal==2.0.0
unicodecsv==0.14.1
update-checker==0.16
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
websockets==6.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
xlwings==0.16.0
xlwt==1.3.0
yarl==1.3.0
zict==1.0.0
zipp==0.6.0
